% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMPNs.R
\name{getMPNs}
\alias{getMPNs}
\title{Get MPNs Function}
\usage{
getMPNs(allData)
}
\arguments{
\item{allData}{The data frame where your data is stored.}
}
\description{
This function allows you to calculate the most probable numbers (MPNs) and their confidence intervals using raw data from an experiment with the IDEXX Quanti-Tray 2000
}
\examples{
getMPNs(allData)

}
\keyword{MPN}
