% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPNdata.R
\docType{data}
\name{MPNdata}
\alias{MPNdata}
\title{Example set of Quanti-Tray 2000 data for multiple samples}
\format{An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.}
\source{
\href{https://safewater.sdsu.edu}{The Safe WaTER Lab at SDSU}
}
\usage{
data(MPNdata)
}
\description{
Data from multiple IDEXX Colilert and Enterolert analyses on water
samples, with one or more dilution levels for each sample. For use
with demonstrations and examples for the MPNcalcalator package.
}
\examples{
data(MPNdata)
}
\references{
Moore et al. (2013) Genetics 195:1077-1086
(\href{https://www.ncbi.nlm.nih.gov/pubmed/23979570}{PubMed})
}
\keyword{datasets}
