% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMPN_prob.R
\name{probScore}
\alias{probScore}
\title{Get MPN Probability Score Function}
\usage{
probScore(mu, dataList)
}
\arguments{
\item{mu}{The MPN value estimated using the getMPN function}

\item{dataList}{The data frame where your data is stored.}
}
\description{
This function uses the binomial distribution to calculate the relative probability of obtaining the observed results of a set of Quanti-Tray 2000 experiments for the same sample at one or more dilution levels (under the assumption that the true concentration is equal to the estimated MPN)
}
\examples{
probScore(dataList)

}
\keyword{MPN}
